requirements.txt
google-api-python-client==2.104.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
pandas==1.5.3
python-dateutil~=2.9.0.post0
httplib2==0.22.0
pip~=25.0.1
protobuf~=6.30.2
pillow~=11.1.0
rsa~=4.9
pyasn1~=0.6.1
pytz~=2025.2
numpy~=2.2.4
charset-normalizer~=3.4.1
setuptools~=78.1.0
pyinstaller~=6.12.0
proto-plus~=1.26.1
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.1.31
urllib3~=2.3.0
packaging~=24.2
requests~=2.32.3
altgraph~=0.17.4
six~=1.17.0
pyparsing~=3.2.3
oauthlib~=3.2.2
idna~=3.10
pefile~=2023.2.7
pywin32-ctypes~=0.2.3
uritemplate~=4.1.1
requests-oauthlib~=2.0.0
pyinstaller-hooks-contrib~=2025.2